Rename Client to Http, use it in Seam module constructor and move factory methods to Seam
